Remove config, create version.py
